requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.4
Django==5.2.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
reportlab==4.4.4
sqlparse==0.5.3
